Reduces database field size from i64 to i32
Optimizes database storage by reducing the field size of several integer types from i64 to i32.

This change reduces the database storage requirements for models like AirBase, PlaneInfo, Hougeki, OpeningTaisen, ClosingRaigeki, OpeningRaigeki, OpeningAirAttack, AirBaseAirAttack, AirBaseAssult, CarrierBaseAssault, SupportHourai, SupportAirattack, FriendlySupportHourai, Battle, Cells, OwnDeck, EnemyShip, FriendShip, OwnShip and OwnSlotItem.

Also introduces a new trait `IntoI32` to handle type conversions from i64 and f32 to i32.

+++ packages/kc_api/crates/kc-api-interface/src/mst_equip_exslot_ship.rs
@@ -25,10 +25,10 @@ pub struct MstEquipExslotShips {
 )]
 #[ts(export, export_to = "get_data.ts")]
 pub struct MstEquipExslotShip {
-    pub ship_ids: Option<HashMap<String, i64>>,
-    pub stypes: Option<HashMap<String, i64>>,
-    pub ctypes: Option<HashMap<String, i64>>,
-    pub req_level: i64,
+    pub ship_ids: Option<HashMap<String, i32>>,
+    pub stypes: Option<HashMap<String, i32>>,
+    pub ctypes: Option<HashMap<String, i32>>,
+    pub req_level: i32,
 }
 
 impl MstEquipExslotShips {
